Trying to change the layout to have 5 rows with 3 columns
