+++ src/components/data.css
@@ -1,0 +1,20 @@
+.Box {
+    flex: 1;
+    align-items: center;
+    align-content: center;
+    background-color: orange;
+    justify-content: center;
+    align-self: center;
+    margin: 50px;
+    border-radius: 25%;
+    width: 200px;
+    height: 200px;
+    flex-wrap: wrap;
+    flex-direction: row
+}
+
+.Picture {
+    height:100px;
+    width:100px;
+    padding-top: 5px;
+}
